services/order-service/requirements-dev.txt
pytest>=8.0
moto[dynamodb]>=5.0
hypothesis>=6.0
boto3-stubs[dynamodb]>=1.34.0
